requirements.txt
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.1.0
certifi==2022.12.7
charset-normalizer==2.1.1
coreapi==2.3.3
coreschema==0.0.4
dj-rest-auth==2.2.5
Django==4.1.4
django-redis==5.2.0
django-rest-swagger==2.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.25.1
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jsonschema==4.17.3
MarkupSafe==2.1.1
openapi-codec==1.3.2
psycopg2-binary==2.9.5
PyJWT==2.6.0
pyrsistent==0.19.2
pytz==2022.6
PyYAML==6.0
redis==4.4.0
requests==2.28.1
simplejson==3.18.0
sqlparse==0.4.3
uritemplate==4.1.1
urllib3==1.26.13
